agent/requirements.txt
supabase
requests
beautifulsoup4
python-dotenv
playwright
